Gracefully deny supplemental transport shutdowns.
This commit ensures that transport shutdowns do not panic on
supplemental shutdowns, even if users should not attempt multiple
shutdowns.  This is done to make the surface for users a little more
forgiving.

The _transport suffix in these implementation filenames are dropped
since they are already part of the transport package, which makes
the specification both redundant and adds stutter.

TEST=``go test ./...``