grpc: add trace.EventLog to ClientConn.
This event log records events that are interesting for live debugging
of a gRPC client.  In particular, this records when a connection is
established, broken, and reestablished.  The log is displayed on the
HTTP endpoint /debug/events in the family named "grpc.ClientConn".

Tested using a version of the route guide client modified to serve
HTTP and block at the end of main.