Update documentation for CGM, in light of jre8u60
JRE 8u60 is still slow, but could be used by someone that doesn't care
about performance now. Added more details about relative performance so
that the gains of OpenSSL are clear.

Added detail about version independence, because JRE 8u60 requires an
update to Jetty ALPN, so it is also applicable to this change.