Allow serviceAccount to reuse auth tokens.
jwt.Config.TokenSource() actually returns a reusable token source,
but serviceAccount actually discards the reusability, it always
creates a new token source instance.

This patch could introduces a redundancy when the token expires;
s.tokenSource.Token() implicitly attempts to obtain the new token
with the old context (the context which s.tokenSource was created
with), and if it fails, a new token source will be created with
a new context and then it attempts to obtain the token again.

I believe this redundancy is ignorable mostly, because this method
falls into one of the following scenarios:
- the same context is used among a program or the current sequence
  of gRPC calls. Then the 'old context' and the 'new context' are
  actually same.
- the old context has enough time by the deadline; simply
  obtaining the new token with the old context succeeds. The new
  context is not used at all.
- the old context is already canceled (or deadline exceeded). Then
  the attempt with the old context fails immediately and switches
  to create a new context.

On the other hand, cancelling the old context or some deadline
settings could lead to the inefficient code path theoretically.
I think that is quite rare though.