Fixes https://github.com/grpc/grpc-go/issues/674.
In the event that we are checking stream quota there is a race condition where we can pick up quota while having a canceled context.  In that case the second wait needs to release the quota.

This does not seem like the correct fix since the issue arises in the wait method -- we only work around it here.