Include codes.DeadlineExceeded code in connection timeout error.
Before this change, some timeout errors returned by gRPC did include
the DeadlineExceeded code. But this was not the case for connection
timeout errors. In order to detect a timeout, an application had to
check for both conditions: a) DeadlineExceeded or b) was the static
connect timeout error returned.

By adding the DeadlineExceeded code to the connect timeout error as well
application code now can be simplified.