proto: avoid deep defensive copy when marshaling
The following pattern is oft-repeated:

    if p := st.Proto(); p != nil {
        bytes, err := proto.Marshal(p)
        ...
    }

Notably it appears in both `transport.WriteStatus` implementations. The
earlier revision with `proto.Clone()` is unnecessary for the immediate
marshaling as the proto message is not mutated. The inner calls to the
`reflect` package shows up on CPU profiles in addition to a tiny
allocation cost that can be easily avoided.

An alternate `Status.ProtoClone()` method is provided for when a deep
copy is in fact necessary, namely when the proto message is the be
subsequently mutated. One such example of this is `Status.WithDetails`
where the embedded proto is annotated with additional detail messages.

NB: This is a breaking API change if imports depend on this specific
behaviour. Alternatively another accessor can be added.