grpc: zero-alloc {Client,Server}Transport.Writes
https://github.com/grpc/grpc-go/pull/1427 avoided the allocation of an
extra buffer by re-using the buffer generated by the Compressor. It did
so by separating out the header and data buffers and delegating to the
transport layer the chunking of the two into appropriately sized frames.

At the transport layer however, given the affinity to generate 16KB
frames, we re-concatenated the header and data buffers and subsequently
flushed out contiguous chunks. Given that the header buffer was always
5KB in size, merging the data bytes into the header buffer necessarily
involved allocations and copying. In practice for every
{Client,Server}Transport.Write we would allocate & copy from 0 to ~16KB
bytes of the data buffer into the header buffer.

By reverting https://github.com/grpc/grpc-go/pull/1427 and re-working
the implementation of rpc_util.go:encode, we can stage a single buffer
comprised of both the serialized{,compressed} message with the
pre-pended message header. The transport layer is again agnostic to the
exact delineation and can simply write out 16KB chunks as needed without
having to copy/allocate.