Make interop client dial blocking (#1559)
This should not be necessary once the new balancer API is done.

Currently interop clients don't use balancer. With a non-blocking dial, the first failfast RPC will fail because there's no connection available.
This is a known bug, and will be fixed by the new balancer APIs.

This is a temp fix to make interop tests pass.