Skip stats test before 1.7
go1.6 doesn't have DialContext (we use Dialer(Cancel: ctx.Done()))
instead. This function sometimes returns non-nil error, while the server
side still gets the conn. This caused the server blocking on reading the
preface (which will eventually timeout, but the tests are too short for
that), and there will be a missing ConnEnd which causes the test to
fail.