Stop retrying on addresses with fatal connecting errors
If connecting to an address returns a non-temporary error, this address will
be removed from the list of addresses that will be retried on.
When there's no more address to retry, that addrConn will stay in
TransientError and then will be torn down (Shutdown).

This was part of #1657 and got reverted to avoid conflict with other changes.

Also it should be OK to call `newAddrConn()` or to update ac.addrs to empty
list. The ac with no connect-able addresses will be torn down.