internal: remove transportMonitor, replace with callbacks
This removes a goroutine (in which transportMonitor was running) and the
backoff and addrs for loops. Instead we use callbacks to handle the
deadline, goaway, and close logic that transport monitor handled
before, and we descend into recursive resetTransport calls (from the
callbacks) instead of loops.

As a result, backoffIdx and addrIdx now live on addrConn. Also,
transport.Close now blocks until the addrConn that initiated it is
connected - this is because t.Close calls onClose, whose last
step is to begin reconnecting. In the event of a shutdown, all
reconnect logic ends and t.Close will return.