auth: Require PRIVACY_AND_INTEGRITY for GoogleCredentials (#4524)
This keeps them more secure. Other types of creds are left as-is, snce
we don't quite know if it makes sense to have a similar restriction. (It
likely does make sense, but this is a more precise change for our
needs.)

This is a rollforward of 8e9d4cbe5cb455315e55846e91706cccecb4b577
which was rolled back in de9515269635d2e8e84540fcd6744cc1a69ce81b

Additional changes that were not in the original commit:
 1. Treat null security level as NONE
 2. When GoogleCredentials can't be loaded, log it at FINE