grpctest: new package
This package makes some of the setup and teardown simpler and more
reliable for test of code that interacts with gRPC servers. It is
non-trivial to write this code in a way that is portable (across OS
types, across supported network interface types, etc.).

This change also demonstrates grpctest's use in the roundrobin test.