internal: ensure http2Client sends END_STREAM
When `http2Client` receives `END_STREAM` from the server, it calls
`closeStream`, but does not send `END_STREAM` back to the server. This
leaves the stream in a half-closed state.

This change ensures that `END_STREAM` is sent back to the server,
putting the stream in a fully closed state. This motivated modifying the
`streamState` type to more closely model the state transitions defined
in https://http2.github.io/http2-spec/#StreamStates. Specifically:
- active -> [recv ES] -> readDone
- active -> [send ES] -> writeDone
- readDone -> [send ES] -> done
- writeDone -> [recv ES] -> done