internal: fix client send preface problems
This CL fixes two problems:

- In clientconn_state_transitions.go, each time we set up a server to send
SETTINGS, we should also set up the server to read. This allows the client
to successfully send its SETTINGS.

- In clientconn.go, we incorrectly transitioned into TRANSIENT FAILURE when
the http2client returned an error. This should be handled in the outer
resetTransport main reset loop. The reason this became a problem is that the
outer resetTransport has very specific conditions around when to transition
into TRANSIENT FAILURE that the egregious transition did not have. So, it
could transition into TRANSIENT FAILURE after failing to dial, even if it
were trying to connect to a non-final address in the list of addresses.