Implementation of server-testing helpers.
End-to-end testing of services with interceptors should be performed with a full
gRPC stack, registered on a Server. This is a convenience package to abstract
the boilerplate code associated with this setup.

Includes a convenience wrapper around creating a grpc.Server communicating over
a bufconn.Listener rather than a loopback TCP connection. The Tester can either
be used directly if multiple services need to be registered, or via the
NewClientConn() function which handles all setup and teardown.