Fix GoAwayThenClose by making RPCs wait-for-ready
In the end of the test, 10 RPCs are made to make sure data is sent to
the second server. The first RPC of these 10 is made right after the
second server's listener receives a connection. But at this time, the
connectivity state on the client side is not set to READY yet. So the
first RPC fails due to transient failure, but the following 9 will
succeed.