examples/helloworld: greeters: add -address
It is now possible to specify an address to bind to on greeter_server
and an address to connect to on greeter_client.

Example:

go run examples/helloworld/greeter_server/main.go -address 192.168.0.20:50052
go run examples/helloworld/greeter_client/main.go -address 192.168.0.20:50052

The flag '-address' is optional; the default is "localhost:50051".