status: Implement *statusError.Is
In the go2 error proposal the Is function adds matching of error targets.
On this line: https://github.com/golang/exp/blob/sumdb/v0.0.2/errors/wrap.go#L58
of the proposal implementation, there is an interface check for Is, to provide
custom implementation of this matching.

gRPC status errors that are defined as a var, currently don't successfully work
with this function when crossing the gRPC boundary, as the pointer values will
not be equivalent.

Ideally, the following should work across gRPC boundaries:
var ErrAuthInvalidToken = status.Error(codes.Unauthenticated, "Invalid token")
_, err := grpcClient.Call(...) // Returns the above
err.Is(ErrAuthInvalidToken) == true