pickfirstBalancer: log messages only at verbose level 2
The pickfirstBalancer is logging a lot of messages under normal
operation.
Those messages can not be associated to a server connection because no
connection address is part of the messages. They are messages that are
only useful when debugging issues.

Only log them when the verbose level is at least 2, to reduce the amount
of log messages under normal operation.