Move xds balancer code from balancer/xds to xds/balancer.
This is to make is possible to add the xds_resolver at xds/resolver. And
later on when the xds_client will be shared by the resolver and
balancer, we could have a separate directory for the client at
xds/client.

Also had to move balancer/internal/wrr to balancer/wrr to be able to
hanve dependencies on that from the xds/balancer code.

Fixed all import paths to make the build and tests happy.