[xds_failover] Support priority
Each priority maps to a priority balancer group, wrapping a priority group.

Priority balancer groups are chained into a linked list, with higher
priority (smaller number) holding pointer to the lower priority (larger
number). When a priority is in use, it's priority balancer group is
started, and it will close it's child priority balancer group (with a
lower priority). When a priority is down (no connection ready), it will
start it's child priority balancer group.