[xds_failover] Support priority
Each priority maps to a priority balancer group, wrapping a priority group.

Priority balancer groups are chained into a linked list, with higher
priority (smaller number) holding pointer to the lower priority (larger
number). When a priority is in use, it's priority balancer group is
started, and it will close it's child priority balancer group (with a
lower priority). When a priority is down (no connection ready), it will
start it's child priority balancer group.

[xds_failover] fix test consts

[xds_failover] refactor to a priority map

[xds_failover] fix race

And also fix a deadlock in test caused by a testing balancer updating
picker while handling addresses.

[xds_failover] renames and comments

[xds_failover] priority lowest

[xds_failover] split handle priority with new state

[xds_failover] Lowest

[xds_failover] priority type

[xds_failover] type more

[xds_failover] split HandleEDSResp

[xds_failover] mutex order

[xds_failover] rename, comments

[xds_failover] timer stop, not reading from channel