Add functionality to read xDS bootstrap file. (#3000)
As of this implementation, the bootstrap file will be provided in the
GRPC_XDS_BOOTSTRAP environment variable and will be read by the xDS
balancer. The file will be in JSON form with two top-level entities: a
Node proto and an ApiConfigSource proto.  The overall JSON format will
be:
{
  "node": <Node proto>,
  "xds_server": <ApiConfigSource proto>
}

In the ApiConfigSource proto, We will support only one grpc_services
entry. The api_type field must be GRPC.

As for creds, we will use default TLS creds (which will use CAs from
system default directories, and will not use any client certificates)
for transport and default compute engine creds for call credentials.
This will evolve into something more configurable in the future.